Use Checkmk notification parameters
Switch configuration from custom contact variables and parameters
to Checkmk parameters that can be edited in WATO. The previous setup
is rather tedious for users and there's always a chance of the
notification rule sending to users that do not have their Telegram
Chat ID set.

This is inspired by the configuration of other push notification types
already shipped with Checkmk, i.e. slack or pushover.